feat: REST reference data can now use dynamic query params (#2539)
refactor: rename graphql variables mapping component to be query params mapping

---------
