+++ libs/shared/src/i18n/en.json
@@ -1859,11 +1859,18 @@
 						"dataSource": {
 							"layoutStyles": {
 								"queryFilters": {
+									"error": "Error while building available reference data variables",
+									"graphql": {
+										"title": "GraphQL variables mapping",
+										"tooltip": "Set values for the GraphQL variables defined in the reference data query. You can get values from the filters using the {{filter.<question-name>}} templates. E.g.: { \"region\": {{filter.region}} }"
+									},
 									"help": "Make sure to remove all unused variables, to prevent null arguments to be sent.",
 									"noVariables": "No variables defined",
-									"refresh": "See all graphQL variables",
-									"title": "GraphQL variable mapping",
-									"tooltip": "Set values for the GraphQL variables defined in the reference data query. You can get values from the filters using the {{filter.<question-name>}} templates. E.g.: { \"region\": {{filter.region}} }"
+									"refresh": "See all variables",
+									"rest": {
+										"title": "REST query params mapping",
+										"tooltip": "Set values for the REST endpoint query params available in the reference data. You can get values from the filters using the {{filter.<question-name>}} templates. E.g.: { \"region\": {{filter.region}} }"
+									}
 								},
 								"title": "Layout styles",
 								"use": "Use layout styles",
