+++ libs/shared/src/i18n/fr.json
@@ -1859,11 +1859,18 @@
 						"dataSource": {
 							"layoutStyles": {
 								"queryFilters": {
+									"error": "Erreur lors de la création des variables de données de référence disponibles",
+									"graphql": {
+										"title": "Mappage de variables GraphQL",
+										"tooltip": "Définissez les valeurs des variables GraphQL définies dans la requête de données de référence. Vous pouvez obtenir les valeurs des filtres à l'aide des modèles {{filter.<question-name>}}. Par exemple: { \"region\": {{filter.region}} }"
+									},
 									"help": "N'oubliez pas de supprimer les variables non utilisées.",
 									"noVariables": "Aucune variable définie",
 									"refresh": "Afficher toutes les variables graphQL",
-									"title": "Mappage de variables GraphQL",
-									"tooltip": "Définissez les valeurs des variables GraphQL définies dans la requête de données de référence. Vous pouvez obtenir les valeurs des filtres à l'aide des modèles {{filter.<question-name>}}. Par exemple : { \"region\": {{filter.region}} }"
+									"rest": {
+										"title": "Mappage de paramètres de requête REST",
+										"tooltip": "Définir les valeurs pour les paramètres de requête du point de terminaison REST disponibles dans les données de référence. Vous pouvez obtenir les valeurs des filtres à l'aide des modèles {{filter.<question-name>}}. Par exemple: { \"region\": {{filter.region}} }"
+									}
 								},
 								"title": "Styles de mise en page",
 								"use": "Utiliser les styles de mise en page",
