+++ libs/shared/src/i18n/test.json
@@ -1859,11 +1859,18 @@
 						"dataSource": {
 							"layoutStyles": {
 								"queryFilters": {
+									"error": "******",
+									"graphql": {
+										"title": "******",
+										"tooltip": "****** {{filter.<question-name>}} ****** { ****** {{filter.region}} ****** }"
+									},
 									"help": "******",
 									"noVariables": "******",
 									"refresh": "******",
-									"title": "******",
-									"tooltip": "****** {{filter.<question-name>}} ****** { ****** {{filter.region}} ****** }"
+									"rest": {
+										"title": "******",
+										"tooltip": "****** {{filter.<question-name>}} ****** { ****** {{filter.region}} ****** }"
+									}
 								},
 								"title": "******",
 								"use": "******",
