+++ libs/shared/src/lib/components/ui/map/map.component.ts
@@ -240,7 +240,7 @@ export class MapComponent
     const allContextFilters = this.layers
       .map((layer: any) => JSON.stringify(layer.contextFilters))
       .join('');
-    const allGraphQLVariables = this.layers
+    const allQueryParams = this.layers
       .map(
         (layer: any) =>
           get(layer, 'datasource.referenceDataVariableMapping') || ''
@@ -249,9 +249,7 @@ export class MapComponent
 
     // Listen to dashboard filters changes to apply layers filter, if it is necessary
     if (
-      this.contextService.filterRegex.test(
-        allContextFilters + allGraphQLVariables
-      )
+      this.contextService.filterRegex.test(allContextFilters + allQueryParams)
     ) {
       this.contextService.filter$
         .pipe(
