+++ libs/shared/src/lib/components/widgets/common/query-params-mapping/query-params-mapping.component.html
@@ -2,8 +2,9 @@
   <span class="flex gap-2 mb-4 items-center">
     <h3 class="mb-0">
       {{
-        'components.widget.settings.summaryCard.card.dataSource.layoutStyles.queryFilters.title'
-          | translate
+        'components.widget.settings.summaryCard.card.dataSource.layoutStyles.queryFilters.' +
+          type +
+          '.title' | translate
       }}
     </h3>
     <ui-icon
@@ -12,11 +13,13 @@
       [size]="18"
       class="cursor-help"
       [uiTooltip]="
-        'components.widget.settings.summaryCard.card.dataSource.layoutStyles.queryFilters.tooltip'
-          | translate
+        'components.widget.settings.summaryCard.card.dataSource.layoutStyles.queryFilters.' +
+          type +
+          '.tooltip' | translate
       "
     ></ui-icon>
     <ui-button
+      *ngIf="type === 'graphql'"
       class="ml-auto"
       category="secondary"
       (click)="refresh(true, true)"
@@ -35,7 +38,7 @@
     'common.errors.invalidJson' | translate
   }}</ui-alert>
   <div
-    *ngIf="availableQueryVariables.length > 0"
+    *ngIf="availableQueryVariables.length > 0 || type === 'rest'"
     mwlResizable
     class="flex flex-col relative h-[300px]"
     [ngStyle]="style"
@@ -55,8 +58,9 @@
       [resizeEdges]="{ bottom: true }"
     ></div>
   </div>
+  <!-- Empty message, when using graphql without any variables -->
   <shared-empty
-    *ngIf="availableQueryVariables.length === 0"
+    *ngIf="type === 'graphql' && availableQueryVariables.length === 0"
     [title]="
       'components.widget.settings.summaryCard.card.dataSource.layoutStyles.queryFilters.noVariables'
         | translate
