+++ libs/shared/src/lib/components/widgets/common/query-params-mapping/query-params-mapping.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ReferenceDataVariablesMappingComponent } from './reference-data-variables-mapping.component';
+
+describe('ReferenceDataVariablesMappingComponent', () => {
+  let component: ReferenceDataVariablesMappingComponent;
+  let fixture: ComponentFixture<ReferenceDataVariablesMappingComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ReferenceDataVariablesMappingComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(ReferenceDataVariablesMappingComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
