+++ libs/shared/src/lib/components/widgets/editor/editor.component.ts
@@ -172,7 +172,7 @@ export class EditorComponent extends BaseWidgetComponent implements OnInit {
     const allContextFilters = this.aggregations
       .map((aggregation: any) => aggregation.contextFilters)
       .join('');
-    const allGraphQLVariables = this.aggregations
+    const allQueryParams = this.aggregations
       .map((aggregation: any) => aggregation.referenceDataVariableMapping)
       .join('');
     // Listen to dashboard filters changes if it is necessary
@@ -193,7 +193,7 @@ export class EditorComponent extends BaseWidgetComponent implements OnInit {
       .subscribe(({ previous, current }) => {
         if (
           this.contextService.filterRegex.test(
-            allContextFilters + allGraphQLVariables
+            allContextFilters + allQueryParams
           )
         ) {
           if (
@@ -421,7 +421,7 @@ export class EditorComponent extends BaseWidgetComponent implements OnInit {
               contextFilters: aggregation.contextFilters
                 ? JSON.parse(aggregation.contextFilters)
                 : {},
-              graphQLVariables: this.widgetService.mapGraphQLVariables(
+              queryParams: this.widgetService.replaceReferenceDataQueryParams(
                 aggregation.referenceDataVariableMapping
               ),
               at: this.contextService.atArgumentValue(aggregation.at),
