+++ libs/shared/src/lib/components/widgets/map-settings/edit-layer-modal/layer-datasource/layer-datasource.component.html
@@ -91,14 +91,14 @@
         ></ui-divider>
       </div>
 
-      <!-- GraphQL variable mapping for reference data graphql-->
-      <shared-graphql-variables-mapping
-        *ngIf="referenceData && referenceData.type === 'graphql'"
+      <!-- Query params mapping for reference data graphql or rest-->
+      <shared-query-params-mapping
+        *ngIf="referenceData && (referenceData.type === 'graphql' || referenceData.type === 'rest')"
         [referenceData]="referenceData"
         [control]="
           $any(formGroup.get('referenceDataVariableMapping'))
         "
-      ></shared-graphql-variables-mapping>
+      ></shared-query-params-mapping>
 
       <!-- Selected layout -->
       <div uiFormFieldDirective *ngIf="layout">
