+++ libs/shared/src/lib/services/aggregation/aggregation.service.ts
@@ -124,7 +124,7 @@ export class AggregationService {
    * @param options.contextFilters context filters, if any
    * @param options.at 'at' argument value, if any
    * @param options.first number of records to fetch, -1 if all of them
-   * @param options.graphQLVariables graphql variables (optional)
+   * @param options.queryParams reference data query params (optional)
    * @returns Aggregation query
    */
   aggregationDataQuery(options: {
@@ -137,7 +137,7 @@ export class AggregationService {
     contextFilters?: CompositeFilterDescriptor;
     at?: Date;
     first?: number;
-    graphQLVariables?: any;
+    queryParams?: any;
   }): Observable<
     ApolloQueryResult<
       AggregationDataQueryResponse | ReferenceDataAggregationQueryResponse
@@ -171,7 +171,7 @@ export class AggregationService {
           contextFilters: options.contextFilters
             ? this.contextService.injectContext(options.contextFilters)
             : {},
-          graphQLVariables: options.graphQLVariables,
+          queryParams: options.queryParams,
           at: options.at,
           first: options.first,
         },
