+++ libs/shared/src/lib/services/aggregation/graphql/queries.ts
@@ -96,7 +96,7 @@ export const GET_REFERENCE_DATA_AGGREGATION_DATA = gql`
     $first: Int
     $skip: Int
     $contextFilters: JSON
-    $graphQLVariables: JSON
+    $queryParams: JSON
     $at: Date
     $sortOrder: String
     $sortField: String
@@ -110,7 +110,7 @@ export const GET_REFERENCE_DATA_AGGREGATION_DATA = gql`
       first: $first
       skip: $skip
       contextFilters: $contextFilters
-      graphQLVariables: $graphQLVariables
+      queryParams: $queryParams
       at: $at
       sortOrder: $sortOrder
       sortField: $sortField
