+++ libs/shared/src/lib/services/widget/widget.service.ts
@@ -187,13 +187,13 @@ export class WidgetService {
   }
 
   /**
-   * Gets graphQLVariables from target source
+   * Gets query params from target source
    *
-   * @param source source we need the mapping variables from
+   * @param source source we need the query params mapping from
    * @param additionalLogic additional logic, for widgets that use other placeholders than filter & context
-   * @returns the graphql query variables object
+   * @returns the reference data query params object
    */
-  public mapGraphQLVariables(
+  public replaceReferenceDataQueryParams(
     source: string | undefined,
     additionalLogic?: (mapping: any) => any
   ): object | null {
