+++ libs/shared/src/lib/survey/components/code-editor/code-editor.component.html
@@ -1,5 +1,5 @@
-<shared-graphql-variables-mapping
+<shared-query-params-mapping
   *ngIf="referenceData && referenceData.type === 'graphql'"
   [referenceData]="referenceData"
   [control]="control"
-></shared-graphql-variables-mapping>
+></shared-query-params-mapping>
